chore(animation): moved animated BallotBox component out of folder
