+++ src/components/Process/BallotBoxAnimated.tsx
